+++ src/execution/PipelineExecutor.ts
@@ -1,6 +1,6 @@
 import type { Promisable } from 'type-fest';
-import type { PipelineJson } from '../types/PipelineJson/PipelineJson';
 import { PipelineExecutionError } from '../errors/PipelineExecutionError';
+import type { PipelineJson } from '../types/PipelineJson/PipelineJson';
 import type { TaskProgress } from '../types/TaskProgress';
 import type { ExecutionReportJson } from '../types/execution-report/ExecutionReportJson';
 import type { Parameters } from '../types/typeAliases';
@@ -11,8 +11,6 @@ import type { PromptResultUsage } from './PromptResultUsage';
  * Executor is made by combining execution tools and pipeline collection.
  *
  * It can be created with `createPipelineExecutor` function.
- *
- * @@@ almost-JSON (what about errors)
  *
  * @see https://github.com/webgptorg/promptbook#executor
  */
@@ -26,7 +24,7 @@ export type PipelineExecutor = {
 /**
  * @@@
  *
- * @@@ almost-JSON (what about errors)
+ * Note: [🚉] This is fully serializable as JSON
  */
 export type PipelineExecutorResult = {
     /**
@@ -49,12 +47,12 @@ export type PipelineExecutorResult = {
     /**
      * Errors that occured during the execution, details are aviable in `executionReport`
      */
-    readonly errors: Array<PipelineExecutionError | Error>;
+    readonly errors: Array<PipelineExecutionError | Error>; // <- !!!!!!
 
     /**
      * Warnings that occured during the execution, details are aviable in `executionReport`
      */
-    readonly warnings: Array<PipelineExecutionError | Error>;
+    readonly warnings: Array<PipelineExecutionError | Error>; // <- !!!!!!
 
     /**
      * The report of the execution with all details
@@ -71,6 +69,7 @@ export type PipelineExecutorResult = {
 };
 
 /**
+ * TODO: [🧠] Should be same array type `errors` and `warnings`
  * TODO: [🧠] Should this file be in /execution or /types folder?
  * TODO: [💷] `assertsExecutionSuccessful` should be the method of `PipelineExecutor` result - BUT maybe NOT?
  */