+++ src/llm-providers/remote/RemoteLlmExecutionTools.ts
@@ -159,14 +159,14 @@ export class RemoteLlmExecutionTools implements LlmExecutionTools {
                 isAnonymous: true,
                 llmToolsConfiguration: this.options.llmToolsConfiguration,
                 prompt,
-                // <- TODO: [🛫] `prompt` is NOT fully serializable as JSON, it contains functions which are not serializable
+                // <- TODO: [🛫] `prompt` is NOT fully [🚉] serializable as JSON, it contains functions which are not serializable
             } satisfies PromptbookServer_Prompt_Request);
         } else {
             socket.emit('prompt-request', {
                 isAnonymous: false,
                 clientId: this.options.clientId,
                 prompt,
-                // <- TODO: [🛫] `prompt` is NOT fully serializable as JSON, it contains functions which are not serializable
+                // <- TODO: [🛫] `prompt` is NOT fully [🚉] serializable as JSON, it contains functions which are not serializable
             } satisfies PromptbookServer_Prompt_Request);
         }