+++ src/types/ModelRequirements.ts
@@ -1,14 +1,11 @@
 import type { ModelVariant } from './ModelVariant';
-import type { number_model_temperature } from './typeAliases';
-import type { number_seed } from './typeAliases';
-import type { string_model_name } from './typeAliases';
-import type { string_system_message } from './typeAliases';
+import type { number_model_temperature, number_seed, string_model_name, string_system_message } from './typeAliases';
 
 /**
  * Abstract way to specify the LLM.
  * It does not specify the LLM with concrete version itself, only the requirements for the LLM.
  *
- * Note: This is fully serializable as JSON
+ * Note: [🚉] This is fully serializable as JSON
  * @see https://github.com/webgptorg/promptbook#model-requirements
  */
 export type ModelRequirements =
@@ -19,7 +16,7 @@ export type ModelRequirements =
 /**
  * Model requirements for the completion variant
  *
- * Note: This is fully serializable as JSON
+ * Note: [🚉] This is fully serializable as JSON
  */
 export type CompletionModelRequirements = CommonModelRequirements & {
     /**
@@ -31,7 +28,7 @@ export type CompletionModelRequirements = CommonModelRequirements & {
 /**
  * Model requirements for the chat variant
  *
- * Note: This is fully serializable as JSON
+ * Note: [🚉] This is fully serializable as JSON
  */
 export type ChatModelRequirements = CommonModelRequirements & {
     /**
@@ -48,7 +45,7 @@ export type ChatModelRequirements = CommonModelRequirements & {
 /**
  * Model requirements for the embedding variant
  *
- * Note: This is fully serializable as JSON
+ * Note: [🚉] This is fully serializable as JSON
  */
 export type EmbeddingModelRequirements = CommonModelRequirements & {
     /**
@@ -62,7 +59,7 @@ export type EmbeddingModelRequirements = CommonModelRequirements & {
 /**
  * Common properties for all model requirements variants
  *
- * Note: This is fully serializable as JSON
+ * Note: [🚉] This is fully serializable as JSON
  */
 export type CommonModelRequirements = {
     /**