+++ src/types/PipelineJson/KnowledgePieceJson.ts
@@ -11,6 +11,8 @@ import type { string_name } from '../typeAliases';
  * Defines one piece of knowledge in the pipeline
  *
  * Note: Knowledge piece is by definition prepared
+ * 
+ * Note: [🚉] This is fully serializable as JSON
  *
  * @see https://github.com/webgptorg/promptbook/discussions/41
  */